refactorize tests to the supported use cases
