+++ tests/_app/api/models/Employee.php
@@ -28,7 +28,7 @@ class Employee extends \app\models\Employee implements Contract
     /**
      * @inheritdoc
      */
-    protected function slugBehaviorConfig()
+    protected function slugBehaviorConfig(): array
     {
         return [
             'resourceName' => 'employee',
