+++ tests/_app/api/models/EmployeeSearch.php
@@ -25,8 +25,10 @@ class EmployeeSearch extends Employee implements \tecnocen\roa\ResourceSearch
     /**
      * @inhertidoc
      */
-    public function search(array $params, $formName = '')
-    {
+    public function search(
+        array $params,
+        ?string $formName = ''
+    ): ?ActiveDataProvider {
         $this->load($params, $formName);
         if (!$this->validate()) {
             return null;
