+++ tests/_app/api/models/Shop.php
@@ -23,7 +23,7 @@ class Shop extends \app\models\Shop implements Contract
     /**
      * @inheritdoc
      */
-    protected function slugBehaviorConfig()
+    protected function slugBehaviorConfig(): array
     {
         return [
             'resourceName' => 'shop',
