--- tests/_app/fixtures/data/file.php
@@ -1,3 +1,0 @@
-<?php
-
-return [];
