--- tests/_app/fixtures/data/restore_soft_delete.php
@@ -1,12 +1,0 @@
-<?php
-
-return [
-    [
-        'id' => 1,
-        'deleted' => true,
-    ],
-    [
-    	'id' => 2,
-    	'deleted' => true,
-    ],
-];
