--- tests/_app/fixtures/data/safe_delete_child.php
@@ -1,8 +1,0 @@
-<?php
-
-return [
-    [
-        'id' => 1,
-        'safe_delete_id'=>2
-    ],
-];
