--- tests/_app/fixtures/data/soft_delete.php
@@ -1,10 +1,0 @@
-<?php
-
-return [
-    [
-        'id' => 1,
-    ],
-    [
-    	'id' => 2,
-    ],
-];
