--- tests/api/RestoreSoftDeleteCest.php
@@ -1,101 +1,0 @@
-<?php
-
-use Codeception\Example;
-use Codeception\Util\HttpCode;
-use app\fixtures\OauthAccessTokensFixture;
-use app\fixtures\RestoreSoftDeleteFixture;
-
-/**
- * Cest to shop resource.
- *
- * @author Carlos (neverabe) Llamosas <carlos@tecnocen.com>
- */
-class RestoreSoftDeleteCest extends \tecnocen\roa\test\AbstractResourceCest
-{
-    protected function authToken(ApiTester $I)
-    {
-        $I->amBearerAuthenticated(OauthAccessTokensFixture::SIMPLE_TOKEN);
-    }
-
-    public function fixtures(ApiTester $I)
-    {
-        $I->haveFixtures([
-            'access_tokens' => OauthAccessTokensFixture::class,
-            'restore_soft_delete' => RestoreSoftDeleteFixture::class,
-        ]);
-    }
-    /**
-     * @param  ApiTester $I
-     * @param  Example $example
-     * @dataprovider indexDataProvider
-     * @depends fixtures
-     * @before authToken
-     */
-    public function index(ApiTester $I, Example $example)
-    {
-        $I->wantTo('Retrieve list of RestoreSoftDelete records.');
-        $this->internalIndex($I, $example);
-    }
-
-    /**
-     * @return array<string,array> for test `index()`.
-     */
-    protected function indexDataProvider()
-    {
-        return [
-            'list' => [
-                'httpCode' => HttpCode::OK,
-                'headers' => [
-                    'X-Pagination-Total-Count' => 2,
-                ],
-            ],
-        ];
-    }
-    /**
-     * @param  ApiTester $I
-     * @param  Example $example
-     * @dataprovider deleteDataProvider
-     * @depends fixtures
-     * @before authToken
-     */
-    public function delete(ApiTester $I, Example $example)
-    {
-        $I->wantTo('Restore a SoftDelete record.');
-        $this->internalDelete($I, $example);
-    }
-    /**
-     * @return array[] data for test `delete()`.
-     */
-    protected function deleteDataProvider()
-    {
-        return [
-            'restore record 1' => [
-                'urlParams' => ['id' => 1],
-                'httpCode' => HttpCode::OK,
-            ],
-            'not found' => [
-                'urlParams' => ['id' => 1],
-                'httpCode' => HttpCode::NOT_FOUND,
-                'validationErrors' => [
-                    'name' => 'The record "1" does not exists.'
-                ],
-            ],
-        ];
-    }
-    /**
-     * @inheritdoc
-     */
-    protected function recordJsonType()
-    {
-        return [
-            'id' => 'integer:>0',
-        ];
-    }
-    /**
-     * @inheritdoc
-     */
-    protected function getRoutePattern()
-    {
-        return 'v1/restore-soft-delete';
-    }
-}
