--- tests/api/SafeDeleteCest.php
@@ -1,103 +1,0 @@
-<?php
-
-use Codeception\Example;
-use Codeception\Util\HttpCode;
-use app\fixtures\OauthAccessTokensFixture;
-use app\fixtures\SafeDeleteFixture;
-use app\fixtures\SafeDeleteChildFixture;
-
-/**
- * Cest to safe-delete resource.
- *
- * @author Carlos (neverabe) Llamosas <carlos@tecnocen.com>
- */
-class SafeDeleteCest extends \tecnocen\roa\test\AbstractResourceCest
-{
-    protected function authToken(ApiTester $I)
-    {
-        $I->amBearerAuthenticated(OauthAccessTokensFixture::SIMPLE_TOKEN);
-    }
-
-    public function fixtures(ApiTester $I)
-    {
-        $I->haveFixtures([
-            'access_tokens' => OauthAccessTokensFixture::class,
-            'safe_delete' => SafeDeleteFixture::class,
-            'safe_delete_child' => SafeDeleteChildFixture::class,
-        ]);
-    }
-    /**
-     * @param  ApiTester $I
-     * @param  Example $example
-     * @dataprovider indexDataProvider
-     * @depends fixtures
-     * @before authToken
-     */
-    public function index(ApiTester $I, Example $example)
-    {
-        $I->wantTo('Retrieve list of SafeDelete records.');
-        $this->internalIndex($I, $example);
-    }
-
-    /**
-     * @return array<string,array> for test `index()`.
-     */
-    protected function indexDataProvider()
-    {
-        return [
-            'list' => [
-                'httpCode' => HttpCode::OK,
-                'headers' => [
-                    'X-Pagination-Total-Count' => 2,
-                ],
-            ],
-        ];
-    }
-    /**
-     * @param  ApiTester $I
-     * @param  Example $example
-     * @dataprovider deleteDataProvider
-     * @depends fixtures
-     * @before authToken
-     */
-    public function delete(ApiTester $I, Example $example)
-    {
-        $I->wantTo('Delete a SafeDelete record.');
-        $this->internalDelete($I, $example);
-    }
-    /**
-     * @return array[] data for test `delete()`.
-     */
-    protected function deleteDataProvider()
-    {
-        return [
-            'soft delete 1' => [
-                'urlParams' => ['id' => 1],
-                'httpCode' => HttpCode::NO_CONTENT,
-            ],
-            'not found' => [
-                'urlParams' => ['id' => 1],
-                'httpCode' => HttpCode::NOT_FOUND,
-                'validationErrors' => [
-                    'name' => 'The record "1" does not exists.'
-                ],
-            ],
-        ];
-    }
-    /**
-     * @inheritdoc
-     */
-    protected function recordJsonType()
-    {
-        return [
-            'id' => 'integer:>0',
-        ];
-    }
-    /**
-     * @inheritdoc
-     */
-    protected function getRoutePattern()
-    {
-        return 'v1/safe-delete';
-    }
-}
