--- tests/api/SoftDeleteCest.php
@@ -1,101 +1,0 @@
-<?php
-
-use Codeception\Example;
-use Codeception\Util\HttpCode;
-use app\fixtures\OauthAccessTokensFixture;
-use app\fixtures\SoftDeleteFixture;
-
-/**
- * Cest to shop resource.
- *
- * @author Carlos (neverabe) Llamosas <carlos@tecnocen.com>
- */
-class SoftDeleteCest extends \tecnocen\roa\test\AbstractResourceCest
-{
-    protected function authToken(ApiTester $I)
-    {
-        $I->amBearerAuthenticated(OauthAccessTokensFixture::SIMPLE_TOKEN);
-    }
-
-    public function fixtures(ApiTester $I)
-    {
-        $I->haveFixtures([
-            'access_tokens' => OauthAccessTokensFixture::class,
-            'soft_delete' => SoftDeleteFixture::class,
-        ]);
-    }
-    /**
-     * @param  ApiTester $I
-     * @param  Example $example
-     * @dataprovider indexDataProvider
-     * @depends fixtures
-     * @before authToken
-     */
-    public function index(ApiTester $I, Example $example)
-    {
-        $I->wantTo('Retrieve list of SoftDelete records.');
-        $this->internalIndex($I, $example);
-    }
-
-    /**
-     * @return array<string,array> for test `index()`.
-     */
-    protected function indexDataProvider()
-    {
-        return [
-            'list' => [
-                'httpCode' => HttpCode::OK,
-                'headers' => [
-                    'X-Pagination-Total-Count' => 2,
-                ],
-            ],
-        ];
-    }
-    /**
-     * @param  ApiTester $I
-     * @param  Example $example
-     * @dataprovider deleteDataProvider
-     * @depends fixtures
-     * @before authToken
-     */
-    public function delete(ApiTester $I, Example $example)
-    {
-        $I->wantTo('Delete a SoftDelete record.');
-        $this->internalDelete($I, $example);
-    }
-    /**
-     * @return array[] data for test `delete()`.
-     */
-    protected function deleteDataProvider()
-    {
-        return [
-            'soft delete 1' => [
-                'urlParams' => ['id' => 1],
-                'httpCode' => HttpCode::NO_CONTENT,
-            ],
-            'not found' => [
-                'urlParams' => ['id' => 1],
-                'httpCode' => HttpCode::NOT_FOUND,
-                'validationErrors' => [
-                    'name' => 'The record "1" does not exists.'
-                ],
-            ],
-        ];
-    }
-    /**
-     * @inheritdoc
-     */
-    protected function recordJsonType()
-    {
-        return [
-            'id' => 'integer:>0',
-        ];
-    }
-    /**
-     * @inheritdoc
-     */
-    protected function getRoutePattern()
-    {
-        return 'v1/soft-delete';
-    }
-}
