+++ tests/api/resources/EmployeeCest.php
@@ -1,17 +1,18 @@
 <?php
 
+namespace resources;
+
+use ApiTester;
+use app\fixtures\OauthAccessTokensFixture;
 use Codeception\Example;
 use Codeception\Util\HttpCode;
-use app\fixtures\EmployeeFixture;
-use app\fixtures\OauthAccessTokensFixture;
-use app\fixtures\ShopFixture;
 
 /**
- * Cest to shop resource.
+ * Cest to Employee resource.
  *
  * @author Carlos (neverabe) Llamosas <carlos@tecnocen.com>
  */
-class ShopCest extends \tecnocen\roa\test\AbstractResourceCest
+class EmployeeCest extends \tecnocen\roa\test\AbstractResourceCest
 {
     protected function authToken(ApiTester $I)
     {
@@ -19,28 +20,15 @@ class ShopCest extends \tecnocen\roa\test\AbstractResourceCest
     }
 
     /**
-     * @depends ProfileCest:fixtures
-     */
-    public function fixtures(ApiTester $I)
-    {
-        $I->haveFixtures([
-            'shop' => ShopFixture::class,
-            'employee' => [
-                'class' => EmployeeFixture::class,
-                'depends' => [],
-            ],
-        ]);
-    }
-    /**
      * @param  ApiTester $I
      * @param  Example $example
      * @dataprovider indexDataProvider
-     * @depends fixtures
+     * @depends resources\ShopCest:fixtures
      * @before authToken
      */
     public function index(ApiTester $I, Example $example)
     {
-        $I->wantTo('Retrieve list of Shop records.');
+        $I->wantTo('Retrieve list of Employee records.');
         $this->internalIndex($I, $example);
     }
 
@@ -51,12 +39,25 @@ class ShopCest extends \tecnocen\roa\test\AbstractResourceCest
     {
         return [
             'list' => [
+                'urlParams' => [
+                    'shop_id' => 2,
+                    'expand' => 'shop'
+                ],
                 'httpCode' => HttpCode::OK,
+                'headers' => [
+                    'X-Pagination-Total-Count' => 2,
+                ],
+            ],
+            'not found shop' => [
+                'urlParams' => [
+                    'shop_id' => 10
+                ],
+                'httpCode' => HttpCode::NOT_FOUND,
             ],
             'filter by name' => [
                 'urlParams' => [
-                    'name' => 'Miniso',
-                    'expand' => 'employees'
+                    'shop_id' => 2,
+                    'name' => 'Robert',
                 ],
                 'httpCode' => HttpCode::OK,
                 'headers' => [
@@ -65,116 +66,142 @@ class ShopCest extends \tecnocen\roa\test\AbstractResourceCest
             ],
         ];
     }
+
     /**
      * @param  ApiTester $I
      * @param  Example $example
      * @dataprovider viewDataProvider
-     * @depends fixtures
+     * @depends resources\ShopCest:fixtures
      * @before authToken
      */
     public function view(ApiTester $I, Example $example)
     {
-        $I->wantTo('Retrieve Shop single record.');
+        $I->wantTo('Retrieve Employee single record.');
         $this->internalView($I, $example);
-        if (isset($example['response'])) {
-            $I->seeResponseContainsJson($example['response']);
-        }
     }
+
     /**
-     * @return array[] data for test `view()`.
+     * @return array<string,array<string,string>> data for test `view()`.
      */
     protected function viewDataProvider()
     {
         return [
-            'expand employees' => [
+            'single record' => [
                 'urlParams' => [
-                    'id' => 1,
-                    'expand' => 'employees'
+                    'shop_id' => 2,
+                    'id' => 3,
+                    'expand' => 'shop'
                 ],
                 'httpCode' => HttpCode::OK,
                 'response' => [
                     '_embedded' => [
-                        'employees' => [
-                            ['id' => 1],
+                        'shop' => [
+                            ['id' => 2],
                         ],
                     ],
                 ],
             ],
+            'not found employee record' => [
+                'url' => '/v1/shop/1/employee/3',
+                'httpCode' => HttpCode::NOT_FOUND,
+            ],
+            'not found shop record' => [
+                'url' => '/v1/shop/10/employee/10',
+                'httpCode' => HttpCode::NOT_FOUND,
+            ],
         ];
     }
+
     /**
      * @param  ApiTester $I
      * @param  Example $example
      * @dataprovider createDataProvider
-     * @depends fixtures
+     * @depends resources\ShopCest:fixtures
      * @before authToken
      */
     public function create(ApiTester $I, Example $example)
     {
-        $I->wantTo('Create a Shop record.');
+        $I->wantTo('Create a Employee record.');
         $this->internalCreate($I, $example);
     }
+
     /**
-     * @return array<string,array<string,array<string,string>>> data for test `create()`.
+     * @return array<string,array<string,string|array<string,string>>> data for test `create()`.
      */
     protected function createDataProvider()
     {
         return [
-            'create shop 4' => [
-                'data' => ['name' => 'Shop 4'],
+            'create employee 3' => [
+                'urlParams' => [
+                    'shop_id' => 1
+                ],
+                'data' => ['name' => 'Employee 3'],
                 'httpCode' => HttpCode::CREATED,
             ],
             'unique' => [
-                'data' => ['name' => 'Shop 4'],
+                'urlParams' => [
+                    'shop_id' => 1
+                ],
+                'data' => ['name' => 'Employee 3'],
                 'httpCode' => HttpCode::UNPROCESSABLE_ENTITY,
                 'validationErrors' => [
-                    'name' => 'Shop Name "Shop 4" has already been taken.'
+                    'name' =>
+                        'Employee Name "Employee 3" has already been taken.'
                 ],
             ],
             'to short' => [
-                'data' => ['name' => 'Shop'],
+                'urlParams' => [
+                    'shop_id' => 1
+                ],
+                'data' => ['name' => 'wo'],
                 'httpCode' => HttpCode::UNPROCESSABLE_ENTITY,
                 'validationErrors' => [
-                    'name' => 'Shop Name should contain at least 6 characters.'
+                    'name' =>
+                        'Employee Name should contain at least 6 characters.'
                 ],
             ],
             'not blank' => [
+                'urlParams' => [
+                    'shop_id' => 1
+                ],
                 'httpCode' => HttpCode::UNPROCESSABLE_ENTITY,
                 'validationErrors' => [
-                    'name' => 'Shop Name cannot be blank.'
+                    'name' => 'Employee Name cannot be blank.'
                 ],
             ],
         ];
     }
+
     /**
      * @param  ApiTester $I
      * @param  Example $example
      * @dataprovider updateDataProvider
-     * @depends fixtures
+     * @depends resources\ShopCest:fixtures
      * @before authToken
      */
     public function update(ApiTester $I, Example $example)
     {
-        $I->wantTo('Update a Shop record.');
+        $I->wantTo('Update a Employee record.');
         $this->internalUpdate($I, $example);
     }
+
     /**
      * @return array[] data for test `update()`.
      */
     protected function updateDataProvider()
     {
         return [
-            'update shop 1' => [
-                'urlParams' => ['id' => '1'],
-                'data' => ['name' => 'Shop 1'],
+            'update employee 1' => [
+                'url' => '/v1/shop/1/employee/1',
+                'data' => ['name' => 'employee 7'],
                 'httpCode' => HttpCode::OK,
             ],
             'to short' => [
-                'urlParams' => ['id' => '1'],
-                'data' => ['name' => 'Shop'],
+                'url' => '/v1/shop/1/employee/1',
+                'data' => ['name' => 'em'],
                 'httpCode' => HttpCode::UNPROCESSABLE_ENTITY,
                 'validationErrors' => [
-                    'name' => 'Shop Name should contain at least 6 characters.'
+                    'name' => 'Employee Name should contain at least 6 characters.'
                 ],
             ],
         ];
@@ -183,33 +210,39 @@ class ShopCest extends \tecnocen\roa\test\AbstractResourceCest
      * @param  ApiTester $I
      * @param  Example $example
      * @dataprovider deleteDataProvider
-     * @depends fixtures
+     * @depends resources\ShopCest:fixtures
      * @before authToken
      */
     public function delete(ApiTester $I, Example $example)
     {
-        $I->wantTo('Delete a Shop record.');
+        $I->wantTo('Delete a Employee record.');
         $this->internalDelete($I, $example);
     }
+
     /**
      * @return array[] data for test `delete()`.
      */
     protected function deleteDataProvider()
     {
         return [
-            'delete shop 1' => [
-                'urlParams' => ['id' => 4],
+            'shop not found' => [
+                'url' => '/v1/shop/10/employee/1',
+                'httpCode' => HttpCode::NOT_FOUND,
+            ],
+            'delete Employee 7' => [
+                'url' => '/v1/shop/1/employee/7',
                 'httpCode' => HttpCode::NO_CONTENT,
             ],
             'not found' => [
-                'urlParams' => ['id' => 4],
+                'url' => '/v1/shop/1/employee/7',
                 'httpCode' => HttpCode::NOT_FOUND,
                 'validationErrors' => [
-                    'name' => 'The record "1" does not exists.'
+                    'name' => 'The record "8" does not exists.'
                 ],
             ],
         ];
     }
+
     /**
      * @inheritdoc
      */
@@ -220,11 +253,12 @@ class ShopCest extends \tecnocen\roa\test\AbstractResourceCest
             'name' => 'string',
         ];
     }
+
     /**
      * @inheritdoc
      */
     protected function getRoutePattern()
     {
-        return 'v1/shop';
+        return 'v1/shop/<shop_id:\d+>/employee';
     }
 }
